Optimize algo: changed dtype from decimal to float in numpy array.

+++ src/core/bitsharesarbitrage.py
@@ -88,7 +88,7 @@ class BitsharesArbitrage(BaseRin):
             order_data_lst = map(
                 lambda order_data: [Decimal(value) for value in order_data.values()], raw_orders_data
             )
-            arr = np.array([*order_data_lst], dtype=Decimal)
+            arr = np.array([*order_data_lst], dtype=float)
 
             try:
                 arr[0]
@@ -100,7 +100,7 @@ class BitsharesArbitrage(BaseRin):
         pairs_orders_data_arrs = await asyncio.gather(
             *(get_order_data_for_pair(pair, market) for pair, market in zip(chain, gram_markets))
         )
-        pairs_orders_data_arr = np.array(pairs_orders_data_arrs, dtype=Decimal)
+        pairs_orders_data_arr = np.array(pairs_orders_data_arrs, dtype=float)
 
         return pairs_orders_data_arr
